Add Windows Support via cross-compilation using xwin for SDKs.
This PR compiles an x86_64 Windows version of the runtime bitcode
on an x86_64 Linux host, using Windows SDKs obtained via
[xwin](https://github.com/Jake-Shadle/xwin).

The resulting runtime has not yet been tested/confirmed to work
on a real Windows machine, but that testing will happen later
after we have updated the Savi compiler to support the Windows target
(and to support cross-compilation more generally, since it's
expected that for now the compiler will not be able to run on Windows
directly, and will have to run within WSL and "cross-compile" to Windows).

This change includes removing the old debuginfo patch and replacing
it with a more general patch that allows us to inject arbitrary CFLAGS,
which are needed for Windows cross-compilation beyond just the
debuginfo.

This patch is expected to be merged into mainline Pony
(see https://github.com/ponylang/ponyc/pull/4108), at which point our
patch here will fail to apply and we'll need to remove it from our repo.