Use LLVM 12 (instead of latest/14) to generate Windows bitcode.
We need the LLVM version we use here to generate the bitcode
to not be newer than the version we'll use in the compiler
to read in the bitcode: 13.0.0 (as set in the `llvm-static` repo).

We specify LLVM 12 instead of LLVM 13 here because the latest
available minor or patch version of LLVM 13 may be newer than
the one we have as our limit (13.0.0).