Add patch to temporarily disable an assert in `ponyint_messageq_destroy`.
This assert gets occasionally violated for the message queue of
the ASIO thread during shutdown. But since it's just shutdown
we don't care much about the implied memory leak.

We eventually want to fix the real race condition but until
then, we'll just take the easy way out and disable the assert.

For continuing investigation, see this Zulip thread:
https://ponylang.zulipchat.com/#narrow/stream/190365-runtime/topic/ponyint_messageq_destroy.20assert.20fail