Fix FreeBSD by removing clobber of system clang.
Prior to this commit, we were installing the `llvm` package
which installs not only LLVM tools but also `clang`, and it
does so by clobbering any existing binaries with those names.

In this commit, we switch to use `llvm13` instead, which only
creates `clang13` and `llvm-link13`. We symlink the latter to
`llvm-link` so we can easily use it, but we leave system `clang`.

We suspect that something with `clang` is causing issues on FreeBSD.

See https://github.com/savi-lang/savi/issues/367